\name{npem-internal}
\alias{avesd.Dsib}
\alias{avesd.Dur}
\alias{choose.rev}
\alias{cluster.stat.sub}
\alias{inherit}
\alias{inherit2}
\alias{Msib}
\alias{Mur}
\alias{parse.hclust}
\alias{parse.hclust.all}
\alias{samppop}

\title{Internal fingers functions}

\description{
  Internal fingers functions.
}

\usage{
avesd.Dsib(p)
avesd.Dur(p)
choose.rev(n)
cluster.stat.sub(x)
inherit(par.g)
inherit2(g,sib)
Msib(p)
Mur(p)
parse.hclust(hclust.out, cutoff)
parse.hclust.all(hclust.out, truefam)
samppop(p)
}

\details{
  These are generally not to be called by the user.
}

\author{Karl W Broman \email{kbroman@biostat.wisc.edu} }

\keyword{internal}
