\name{shiff1}
\alias{shiff1}

\title{Schistosome data}
\description{ 
  This is RAPD data for 35 loci typed on a set of 135 individuals.
} 
\usage{data(shiff1)}
\format{
  The data is a matrix of 135 rows (the individuals) by 35 columns (the
  RAPD loci).  Each entry is a RAPD phenotype, indicating the presence
  (1) or absence (0) of a band.
}

\source{Clive Shiff, Molecular Microbiology and Immunology,
  Bloomberg School of Public Health, The Johns Hopkins University
}

\author{Karl W Broman \email{kbroman@biostat.wisc.edu} }

\examples{
data(shiff1)
}

\seealso{\code{\link{shiff2}}, \code{\link{shiff3}},
  \code{\link{aedes}}, \code{\link{simrapd}} }

\keyword{datasets}
